+++ tests/packages/test_locker.py
@@ -61,3 +61,31 @@ B = []
 """
 
     assert expected == content
+
+
+def test_lock_packages_with_null_description(locker, root):
+    package_a = get_package("A", "1.0.0")
+    package_a.description = None
+
+    locker.set_lock_data(root, [package_a])
+
+    with locker.lock.open(encoding="utf-8") as f:
+        content = f.read()
+
+    expected = """[[package]]
+category = "main"
+description = ""
+name = "A"
+optional = false
+python-versions = "*"
+version = "1.0.0"
+
+[metadata]
+content-hash = "115cf985d932e9bf5f540555bbdd75decbb62cac81e399375fc19f6277f8c1d8"
+python-versions = "*"
+
+[metadata.hashes]
+A = []
+"""
+
+    assert expected == content